Move reconciler out of app/, split DNS from Route53 code + test
