+++ test/unit/route53.server.test.ts
@@ -1,7 +1,4 @@
-import {
-  toRoute53RecordValue,
-  fromRoute53RecordValue,
-} from '~/queues/reconciler/route53Utils.server';
+import { toRoute53RecordValue, fromRoute53RecordValue } from '../../reconciler/route53Utils.server';
 import { DnsRecordType } from '@prisma/client';
 
 describe('Route53 functionality test raw => Route53 format', () => {
